Make CF ignorance it's own thing;
This used to be okay for cloud.gov EW to not have to vendor the
dependencies, but we run out of memory on GovCloud if we deploy NodeJS
applications without vendoring dependencies. So we need to not just
symlink the `gitignore` file but copy it and remove the `node_modules/`
line in order to make sure it gets uploaded on `cf push`.

This change is going to end up using a lot more bandwidth when it comes
to deploying this application.