Session expiration warning.  (#1649)
* Request CSRF token when prompting user to log in.

* Session expiration warning.

Added an idle modal that alert the user that due to inactivity they will
be logged out. The warning give the user a three minute window in which
to reset his/her session. Sessions are currently set for 30 minutes,
therefore the user will be alerted after 27 minutes of inactivity.

Reset session timer on page navigation. Reset session timer on 30
keypresses or one minute lapse in time since last keypress.